+++ main.d.ts
@@ -7,7 +7,7 @@
  * @example
  * ```bash
  * # On the VPS
- * tunnelr -p 20185
+ * tunnelr server -p 20185
  *
  * # On your machine
  * tunnelr my-vps.example.com:20185 -p 30185:3000